requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-tools==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.7.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
requests==2.32.2
requests-oauthlib==2.0.0
robotframework==7.0
robotframework-assertion-engine==3.0.3
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.3.0
rsa==4.9
selenium==4.20.0
setuptools==69.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
wsproto==1.2.0
